feat: added seed data from old web version and fixed some db types

+++ lib/data/models/ui/ingredient_input.dart
@@ -30,7 +30,7 @@ class IngredientInput {
     return RecipeIngredientsCompanion.insert(
       recipeId: recipeId,
       pantryId: pantryId,
-      quantityNeeded: int.tryParse(qtyController.text) ?? 0,
+      quantityNeeded: double.tryParse(qtyController.text) ?? 0.0,
       unit: unitController.text,
     );
   }
